requirements.txt
discord.py
yt-dlp
googletrans
Pillow
openai
transformers
requests
python-dotenv